+++ OSXvnc-server/rdr/NullOutStream.cxx
@@ -45,7 +45,7 @@ void NullOutStream::writeBytes(const void* data, size_t length)
   offset += length;
 }
 
-size_t NullOutStream::overrun(int itemSize, size_t nItems)
+size_t NullOutStream::overrun(size_t itemSize, size_t nItems)
 {
   if (itemSize > bufferSize)
     throw Exception("NullOutStream overrun: max itemSize exceeded");
